+++ packages/cli/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - Refactored `moon init` heavily for a better onboarding experience.
   - Each tool is now configured individually, with its own prompts. Tools can also be skipped.
-  - Tools can now be initialized _after_ moon already exists, ala `moon init node`.
+  - Tools can now be initialized _after_ moon already exists, ala `moon init --tool node`.
   - Fixed many issues around the templates and rendering.
 - Updated `moon migrate` commands to throw an error if the work tree is dirty. This can be bypassed
   with the new `--skipTouchedFilesCheck` option.